[Refactor] camera driver Now, camera driver is an abstract class and need subclass as concrete class to perform concrete method. By doing this, we separate camera (base class) and usb camera(sub class), so we enhance loose coupling between objects.
Also, we defined an interface for camera. By doing this, we provide stable
driver API for users.

[Refactor] test_camera
Refactored test_camera's tests cases. We made it concise, readable, and
making appropriate tests.

[Refactor] camera_ex.py

Refactored the syntaxs. We made it readable and folows the new camera
driver method.

[Delete] pyc files and test_usb_camera.py
Pycs are automatically built when we run the program (it doesn't affect
the major program).

We deleted the old tests case program and replace it with the camera test.py.
The old tests case was built in cumbersome structure and inappropriate tests case.